testing fixes, logic removed from member view
