+++ app/models/event.rb
@@ -5,23 +5,31 @@ class Event < ApplicationRecord
     has_many :due, dependent: :destroy
     validates :name, :description, :location, :start_time, :end_time, presence: true
     validate :start_greater_equal, :inFuture
-
+    
     def start_greater_equal
+
         if start_time == end_time
             errors.add(:start_time, 'must be before end time')
             errors.add(:end_time, 'must be after start time')
-        end
-        if start_time > end_time
+        elsif start_time == nil
+            errors.add('requires start time')
+        elsif end_time == nil
+            errors.add('requires end time')
+        elsif start_time > end_time
             errors.add(:start_time, 'must be before end time')
             errors.add(:end_time, 'must be after start time')
         end
+        
     end
 
     def inFuture
-        if start_time < DateTime.current-6.hours
+        if start_time == nil
+            errors.add('requires start time')
+        elsif end_time == nil
+            errors.add('requires end time')
+        elsif start_time < DateTime.current-6.hours
             errors.add(:start_time, 'must not have already passed')
-        end
-        if end_time < DateTime.current-6.hours
+        elsif end_time < DateTime.current-6.hours
             errors.add(:end_time, 'must not have already passed')
         end
     end
