+++ app/views/members/index.html.erb
@@ -18,7 +18,7 @@
               <th>Major</th>
               <th>Active</th>
               <%# Hidden to Members %>
-              <% if Member.find_by(:email =>current_admin.email).admin? %>
+              <% if @admin %>
                 <th>Phone Number</th>
                 <th>Admin</th>
                 <th colspan="3"></th>
@@ -37,7 +37,7 @@
                 <td><%= member.major %></td>
                 <td><%= member.active %></td>
                 <%# Hidden to Members %>
-                <% if Member.find_by(:email =>current_admin.email).admin? %>
+                <% if @admin %>
                   <td><%= member.phoneNumber %></td>
                   <td><%= member.admin %></td>
                   <td><%= link_to 'Show', member %></td>
